requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
pip==23.2.1
psycopg2-binary==2.9.9
requests==2.32.3
setuptools==68.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.41.2
